MWPW-189894: Rename defaultLanguageVariation -> defaultLanguage
