+++ io/www/src/fragment/pipeline.js
@@ -6,7 +6,7 @@ import zlib from 'zlib';
 import stateLib from '@adobe/aio-lib-state';
 
 import { transformer as fetchFragment } from './transformers/fetchFragment.js';
-import { transformer as defaultLanguageVariation } from './transformers/defaultLanguageVariation.js';
+import { transformer as defaultLanguage } from './transformers/defaultLanguage.js';
 import { transformer as corrector } from './transformers/corrector.js';
 import { transformer as replace } from './transformers/replace.js';
 import { transformer as promotions } from './transformers/promotions.js';
@@ -22,7 +22,7 @@ function calculateHash(body) {
     return crypto.createHash('sha256').update(JSON.stringify(body)).digest('hex');
 }
 
-const PIPELINE = [fetchFragment, defaultLanguageVariation, promotions, customize, settings, replace, wcs, corrector];
+const PIPELINE = [fetchFragment, defaultLanguage, promotions, customize, settings, replace, wcs, corrector];
 
 const RESPONSE_HEADERS = {
     'Access-Control-Expose-Headers': 'X-Request-Id,Etag,Last-Modified,server-timing',
